Fixed gamepad controls & money bag bug
Although initial gamepad support was added for the in-level controls, you couldn't actually navigate the titlescreen or gameover screen with a controller. That has been fixed, and now the game is fully playable with a controller. Also, the bug where the  money bag scaling would get all out of whack if you were collecting coins quickly is fixed. The scale now resets with every coin.