Implemented enemy damage
So, it took a bit of screwing around, but I cleaned up a lot of things, and got enemy damage working. So...now if an enemy has 2 health, it has to take 2 damage to be killed. I find it odd that the Signals that Godot uses aren't unique to all instances of a scene. I was having a hard time trying to find out why hitting one enemy would kill all of the enemies of the same type that are on screen. Turns out, I had to implement a Guid per enemy to check that the correct one was being hit. I'll make this into an Interface in the future so all enemies can implement it.