Quality of Life
Added a bunch of QoL changes that get us more towards demo ready. First, added music to the credits screen, got rid of assets we don't need, and added drop-shadows to the HUD asset. Really makes the game pop a bit. Also added a pause button to the in-game behavior, also with a drop shadow. Additionally, when you get hit, the hearts fall down and disappear, rather than the old harsh transition. Also...when you die, the gameplay slowly slows down, and the screen fades to black.

Slowly working through the punch-list to get to the demo. I have Lennon's drawing for the level progress bar, boss health bar, and helpers. Just need to scan them in and code them into the game. Also, I owe drawings for the story at the start of the game, and a few other things.