Boss fight!
I cleaned up a ton of un-used resources, scanned in and added some new graphics (mutant fly!) and started the implementation of the boss at the end of level 1. The whole "movement" thing really needs a refactor into the enemy base class as well as a million more cleanups, but no time for that now.