Boss nearly done
I worked out a number of bugs in the boss fight, so now (minus some difficulty tweaks) the fight is complete. I do need to animated the boss's mouth when it shoots out a fly, but that should be easy to do. Only a few more things left on the punch list before the demo is done.