Boss finished!
All the nuts and bolts of the boss fight are implemented and working. I really need to probably do some tweaking to make it hard but not too challenging, but hopefully that's it. 

I plan on tackling difficulty by moving a number of important variables into the Globals class to make them centrally controllable. Things like player rotation speed, spawn rates, enemy movement speed, etc. We'll see. I'm really hoping to have a rough cut of the demo finished by the end of the day.