Demo - feature complete!
So...with a bunch of last minute fiddling it appears that I at least have all of the main features I wanted to add to the demo in there. Obviously there is a bit of polish and a few things left to do, but you can play from start to finish for level 1 at this point.