Difficulty selection
I was able to scan in the assets, and then implement a difficulty selection on the titlescreen. So far, the initial values are just guesses and haven't been thoroughly play tested, but I can handle that in the future. 

There may be a bug in hard difficulty mode that causes a crash during level play, but I haven't had time to dig into that either.