Complete!
So...it looks like the demo is complete. I still need to do some more cleanup, icons, etc...but for now, it's complete. You can play from start to finish with cutscenes, credits, etc. It's all there.

I want to also add a bit of polish, but we'll see how motivated I feel.